# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0", 
        "title": "My Library Project CRUD",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
    "tags": [
        {
            "name": "Library",
            "description": "API for users in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],

    "paths": {
        "/books": {
            "get": {
                "tags": ["Books"],
                "summary": "Get all Books in library",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                }
            }
        },
        "/book": {
            "post": {
                "tags": ["Books"],
                "description": "Create new book in library",
                "parameters": [
                    {
                        "name": "book",
                        "in": "body",
                        "description": "Book that we want to create",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "New book is created",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                }
            }
        },
        "/book/{id}": {
            "delete": {
                "tags": ["Books"],
                "description": "Delete book in library",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of book that we want to delete",
                        "type": "integer",
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book is deleted",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Book": {
            "required": ["id","name","author"],
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                }
            }
        }
    }
}